Directory file provider takes sub-directories into account :)

+++ app/features/utils/lib/clean.js
@@ -1,13 +1,24 @@
 var fs = require('fs');
 var array = require('./array.utils');
 
-module.exports = {
+var clean = {
 	
 	folder: function(folder) {
 		if (!fs.existsSync(folder)) fs.mkdirSync(folder);
 		var files = fs.readdirSync(folder);
 		array.forEach(files, function(file) {
-			fs.unlinkSync(folder + '/' + file);
+			
+			var stats = fs.statSync(folder + '/' + file);
+			
+			if (stats.isFile()) {
+				fs.unlinkSync(folder + '/' + file);
+			}			
+			if (stats.isDirectory()) {
+				clean.folder(folder + '/' + file);
+				fs.rmdirSync(folder + '/' + file);
+			}
 		});		
 	},
 };
+
+module.exports = clean;
